+++ cpp/subprojects/common/test/mlrl/common/input/feature_vector_decorator_numerical.cpp
@@ -74,7 +74,7 @@ TEST(NumericalFeatureVectorDecoratorTest, updateCoverageMaskAndStatisticsInverse
     CoverageMask coverageMask(numExamples);
     uint32 indicatorValue = 1;
     decorator.updateCoverageMaskAndStatistics(interval, coverageMask, indicatorValue, statistics);
-    EXPECT_EQ(coverageMask.getIndicatorValue(), 0);
+    EXPECT_EQ(coverageMask.getIndicatorValue(), (uint32) 0);
 
     for (uint32 i = 0; i < interval.start; i++) {
         EXPECT_TRUE(coverageMask.isCovered(i));
@@ -164,7 +164,7 @@ TEST(NumericalFeatureVectorDecoratorTest, updateCoverageMaskAndStatisticsFromVie
     CoverageMask coverageMask(numDenseExamples);
     uint32 indicatorValue = 1;
     filtered->updateCoverageMaskAndStatistics(interval, coverageMask, indicatorValue, statistics);
-    EXPECT_EQ(coverageMask.getIndicatorValue(), 0);
+    EXPECT_EQ(coverageMask.getIndicatorValue(), (uint32) 0);
 
     for (uint32 i = 0; i < interval.start; i++) {
         EXPECT_TRUE(coverageMask.isCovered(i));
